+++ modules/api/current.js
@@ -121,6 +121,7 @@ module.exports = {
     const { id, testCaseId, attempt } = data;
     this.storage.setTestCaseStartedId(id, testCaseId);
     const { pickleId, isRetry: isTestCaseRetried } = this.storage.getTestCase(testCaseId);
+
     const {
       uri: pickleFeatureUri,
       astNodeIds: [scenarioId, parametersId],
@@ -128,11 +129,31 @@ module.exports = {
     const currentFeatureUri = this.storage.getCurrentFeatureUri();
     const feature = this.storage.getFeature(pickleFeatureUri);
     const launchTempId = this.storage.getLaunchTempId();
-
+    const isFirstFeatureInLaunch = !currentFeatureUri;
+    const isNewFeature = currentFeatureUri !== pickleFeatureUri;
     // start FEATURE if no currentFeatureUri or new feature
     // else finish old one
     const featureCodeRef = utils.formatCodeRef(pickleFeatureUri, feature.name);
-    if (currentFeatureUri !== pickleFeatureUri) {
+    if (isFirstFeatureInLaunch) {
+      this.storage.setCurrentFeatureUri(pickleFeatureUri);
+
+      const suiteData = {
+        name: feature.name,
+        startTime: this.reportportal.helpers.now(),
+        type: this.isScenarioBasedStatistics ? TEST_ITEM_TYPES.TEST : TEST_ITEM_TYPES.SUITE,
+        description: (feature.description || '').trim(),
+        attributes: utils.createAttributes(feature.tags),
+        codeRef: featureCodeRef,
+      };
+
+      const { tempId } = this.reportportal.startTestItem(suiteData, launchTempId, '');
+      this.storage.setFeatureTempId(tempId);
+    } else if (isNewFeature) {
+      const previousFeatureTempId = this.storage.getFeatureTempId();
+      this.reportportal.finishTestItem(previousFeatureTempId, {
+        endTime: this.reportportal.helpers.now(),
+      });
+
       this.storage.setCurrentFeatureUri(pickleFeatureUri);
       const suiteData = {
         name: feature.name,
@@ -144,14 +165,10 @@ module.exports = {
       };
       const { tempId } = this.reportportal.startTestItem(suiteData, launchTempId, '');
       this.storage.setFeatureTempId(tempId);
-    } else {
-      const tempFeatureId = this.storage.getFeatureTempId();
-      this.reportportal.finishTestItem(tempFeatureId, {
-        endTime: this.reportportal.helpers.now(),
-      });
     }
 
-    // current feature node rule || scenario
+    // current feature node rule(this entity is for grouping several
+    // scenarios in one logical block) || scenario
     const currentNode = utils.findNode(feature, scenarioId);
 
     let scenario;
@@ -444,27 +461,21 @@ module.exports = {
 
     this.storage.setStepTempId(null);
   },
-  onTestCaseFinishedEvent(data) {
-    const { testCaseStartedId, willBeRetried } = data;
+  onTestCaseFinishedEvent({ testCaseStartedId, willBeRetried }) {
+    const isNeedToFinishTestCase = !this.isScenarioBasedStatistics && willBeRetried;
+
+    if (isNeedToFinishTestCase) {
+      return;
+    }
+
     const testCaseId = this.storage.getTestCaseId(testCaseStartedId);
     const testCase = this.storage.getTestCase(testCaseId);
     const scenarioTempId = this.storage.getScenarioTempId();
+
     this.reportportal.finishTestItem(scenarioTempId, {
       endTime: this.reportportal.helpers.now(),
       ...(this.isScenarioBasedStatistics && { status: testCase.status || STATUSES.PASSED }),
     });
-
-    const currentFeatureUri = this.storage.getCurrentFeatureUri();
-    const featureHasOneTestCase =
-      Array.from(this.storage.getPicklesValues()).filter(({ uri }) => uri === currentFeatureUri)
-        .length === 1;
-
-    if (featureHasOneTestCase) {
-      const tempFeatureId = this.storage.getFeatureTempId();
-      this.reportportal.finishTestItem(tempFeatureId, {
-        endTime: this.reportportal.helpers.now(),
-      });
-    }
 
     // finish RULE if it's exist and if it's last scenario
     const isLastScenario = this.storage.getLastScenario();